+++ R/export_table.R
@@ -25,12 +25,17 @@
 #' @param title,caption,subtitle Table title (same as caption) and subtitle, as
 #'   strings. If `NULL`, no title or subtitle is printed, unless it is stored as
 #'   attributes (`table_title`, or its alias `table_caption`, and
-#'   `table_subtitle`). If `x` is a list of data frames, `caption` may be a list
-#'   of table captions, one for each table.
+#'   `table_subtitle`). If you want to force that no title is printed, even if
+#'   present as attribute, use `""`, which will never print titles. If `x` is a
+#'   list of data frames, `caption` may be a list of table captions, one for
+#'   each table.
 #' @param footer Table footer, as string. For markdown-formatted tables, table
 #'   footers, due to the limitation in markdown rendering, are actually just a
 #'   new text line under the table. If `x` is a list of data frames, `footer`
-#'   may be a list of table captions, one for each table.
+#'   may be a list of table captions, one for each table. If `NULL`, no footer
+#'   is printed, unless it is stored as attributes (`table_footer`). If you want
+#'   to force that no footer is printed, even if present as attribute, use `""`,
+#'   which will never print footers.
 #' @param align Column alignment. For markdown-formatted tables, the default
 #'   `align = NULL` will right-align numeric columns, while all other columns
 #'   will be left-aligned. If `format = "html"`, the default is left-align first
@@ -516,6 +521,17 @@ print.insight_table <- function(x, ...) {
                           verbose = TRUE,
                           ...) {
   table_data <- as.data.frame(x)
+
+  # we set empty caption, subtitle and footer to NULL, so they are not printed
+  if (identical(caption, "")) {
+    caption <- NULL
+  }
+  if (identical(subtitle, "")) {
+    subtitle <- NULL
+  }
+  if (identical(footer, "")) {
+    footer <- NULL
+  }
 
   # rename columns?
   table_data <- .new_column_names(table_data, column_names)
@@ -1278,7 +1294,7 @@ print.insight_table <- function(x, ...) {
   n_columns <- ncol(final)
   first_row_leftalign <- (!is.null(align) && align == "firstleft")
 
-  ## create header line for markdown table -----
+  # create header line for markdown table -----
   header <- "|"
 
   # indent groups?
@@ -1425,7 +1441,7 @@ print.insight_table <- function(x, ...) {
   ))
 
   # validation check - clean caption, subtitle and footer from ansi-colour codes,
-  # which only work for text format... But if user occidentally provides colours
+  # which only work for text format... But if user accidentally provides colours
   # for HTML format as well, remove those colour codes, so they don't appear as
   # text in the table header and footer. Furthermore, in footers, we need to
   # remove newline-characters
@@ -1462,7 +1478,11 @@ print.insight_table <- function(x, ...) {
 
   tab <- gt::gt(final, groupname_col = group_by_columns)
   header <- gt::tab_header(tab, title = caption, subtitle = subtitle)
-  footer <- gt::tab_source_note(header, source_note = gt::html(footer))
+  if (!is.null(footer)) {
+    footer <- gt::tab_source_note(header, source_note = gt::html(footer))
+  } else {
+    footer <- gt::tab_source_note(header, source_note = NULL)
+  }
   out <- gt::cols_align(footer, align = "center")
 
   # emphasize header of row groups?