+++ tests/testthat/test-export_table.R
@@ -366,3 +366,26 @@ test_that("export_table, tinytable with indented rows", {
   )
   expect_snapshot(export_table(mp, format = "tt", by = "groups", table_width = Inf))
 })
+
+
+test_that("export_table, removing captions work", {
+  skip_on_cran()
+  skip_if_not_installed("modelbased")
+  skip_if_not_installed("marginaleffects")
+
+  data(iris)
+  mod <- lm(Petal.Length ~ Species, data = iris)
+  means <- modelbased::estimate_means(mod, by = "Species", type = "response")
+
+  expect_snapshot(print(means, table_width = Inf))
+  expect_snapshot(print(means, title = "", table_width = Inf))
+  expect_snapshot(print(means, caption = "", table_width = Inf))
+  expect_snapshot(print(means, footer = "", caption = "", table_width = Inf))
+
+  skip_if_not_installed("gt")
+  set.seed(123)
+  out <- gt::as_raw_html(print_html(means))
+  expect_snapshot(as.character(out))
+  out <- gt::as_raw_html(print_html(means, footer = "", caption = ""))
+  expect_snapshot(as.character(out))
+})